Improvements on Profile Activity
Now details from the user is saved into the DB powered by Firebase (In a future will be BigchainDB)

Job role big has been fixed. Now when a user saves the job role is displayed and saven into the DB.

It needs to read the data from the DB when starting the application.